+++ activerecord/lib/active_record/connection_adapters/mysql/schema_creation.rb
@@ -49,6 +49,7 @@ module ActiveRecord
             sql << "USING #{o.using}" if o.using
             sql << "ON #{quote_table_name(o.table)}" if create
             sql << "(#{quoted_columns(o)})"
+            sql << "WITH PARSER #{o.with_parser}" if o.with_parser
 
             add_sql_comment!(sql.join(" "), o.comment)
           end